backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-chartjs==2.3.0
django-cors-headers==4.4.0
django-material-admin==1.8.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
